+++ strengthjournal-server/StrengthJournal.Server/ApiModels/WorkoutDto.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; set; }
         public string Title { get; set; }
         public DateTime EntryDateUTC { get; set; }
-        public uint? Bodyweight { get; set; }
+        public decimal? Bodyweight { get; set; }
         public string BodyweightUnit { get; set; }
         public IEnumerable<WorkoutSetSync> Sets { get; set; }
     }
